requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
boto=2.49.0=pypi_0
boto3=1.12.7=pypi_0
botocore=1.15.7=pypi_0
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_0
chardet=3.0.4=pypi_0
cycler=0.10.0=pypi_0
docutils=0.15.2=pypi_0
gensim=3.8.1=pypi_0
idna=2.9=pypi_0
jieba=0.42.1=pypi_0
jmespath=0.9.5=pypi_0
joblib=0.14.1=pypi_0
kiwisolver=1.1.0=pypi_0
lxml=4.5.0=pypi_0
matplotlib=3.1.3=pypi_0
numpy=1.18.1=pypi_0
openssl=1.1.1d=he774522_4
pandas=1.0.1=pypi_0
pillow=7.0.0=pypi_0
pip=20.0.2=py37_1
pyparsing=2.4.6=pypi_0
python=3.7.6=h60c2a47_2
python-dateutil=2.8.1=pypi_0
pytz=2019.3=pypi_0
requests=2.23.0=pypi_0
s3transfer=0.3.3=pypi_0
scikit-learn=0.22.1=pypi_0
scipy=1.4.1=pypi_0
setuptools=45.2.0=py37_0
six=1.14.0=pypi_0
sklearn=0.0=pypi_0
smart-open=1.9.0=pypi_0
urllib3=1.25.8=pypi_0
wheel=0.34.2=py37_0
wincertstore=0.2=py37_0
wordcloud=1.6.0=pypi_0
yapf=0.28.0=py_0
